DS-308-rich_text_component_fixes
